Collapse foreman core namespace
