--- backend/foreman_core/__init__.py
@@ -1,1 +1,0 @@
-"""Shared core library for Pioneer Square foreman (embedded + standalone)."""
