+++ cli/pioneer_cli/main.py
@@ -62,7 +62,7 @@ def _run_serve(argv: list[str]) -> int:
         return 2
 
     # The backend imports its modules as top-level names (``from events import``,
-    # ``main:app``, ``foreman_core``...) and resolves Alembic/static paths from
+    # ``main:app``, ``foreman``...) and resolves Alembic/static paths from
     # its own directory, so run it exactly as ``python main.py`` would.
     _prepend_path(backend_dir)
     os.chdir(backend_dir)
@@ -95,8 +95,8 @@ def _run_worker(argv: list[str]) -> int:
 def _run_foreman(argv: list[str]) -> int:
     """Launch the standalone Foreman API proxy (delegates to pioneer_foreman.cli.main)."""
     root = _resolve_root()
-    # foreman/ holds the pioneer_foreman package; root makes ``backend.foreman_core``
-    # importable from source (backend/__init__.py + backend/foreman_core/).
+    # foreman/ holds the pioneer_foreman package; root makes ``backend.foreman.llm``
+    # importable from source without importing the backend runtime.
     _prepend_path(root / "foreman")
     _prepend_path(root)
     from pioneer_foreman.cli import main as foreman_main  # noqa: PLC0415
